Self-casting Pot().

+++ klib.h
@@ -1,6 +1,6 @@
 int nascii(char); // numeri decimali dai rispettivi caratteri
 int qualcosa(int,char*,char*); // (argc-n (n= quante cose), "quali cose",argv[0])
-long pot(int,int); // pot(x,y)= x^y
+long pot(long,int); // pot(x,y)= x^y
 long unit(int*,int); // rende un intero da un array (arg1) di n (arg2) numeri
 long deci(char*,int); // rende un intero da un array (arg1) di n (arg2) caratteri
-size_t strl(char*); // strl( stringa )= lunghazza della stringa
+size_t strl(char*); // strl("stringa")= lunghazza della stringa
